requirements.txt
selenium==4.9.1
robotframework-seleniumlibrary
robotframework
addict
robotframework-pabot